declare loop variable i as unsigned int

+++ src/test/OpenEXRTest/testDwaLookups.cpp
@@ -226,9 +226,9 @@ testNoop ()
 {
     printf ("test dwaCompressorNoOp[] \n");
 
-    for (int i = 0; i < 65536; ++i)
+    for (unsigned int i = 0; i < 65536; ++i)
     {
-        assert ((unsigned short) i == OPENEXR_IMF_INTERNAL_NAMESPACE::dwaCompressorNoOp[i]);
+        assert (i == OPENEXR_IMF_INTERNAL_NAMESPACE::dwaCompressorNoOp[i]);
     }
 }
